Reorganise folders/namespaces
