--- TRDataControl/Events/SegmentEventArgs.cs
@@ -1,9 +1,0 @@
-﻿using TRImageControl;
-
-namespace TRModelTransporter.Events;
-
-public class SegmentEventArgs : EventArgs
-{
-    public int SegmentIndex { get; set; }
-    public TRImage Image { get; set; }
-}
