+++ TRDataControl/Transport/TR1/TR1TextureRemapper.cs
@@ -1,7 +1,7 @@
 ﻿using TRImageControl.Packing;
 using TRLevelControl.Model;
 
-namespace TRDataControl.Remapping;
+namespace TRDataControl;
 
 public class TR1TextureRemapper : TRTextureRemapper<TR1Level>
 {
