+++ TRDataControl/Transport/TR2/TR2DataExporter.cs
@@ -1,6 +1,5 @@
 ﻿using System.Diagnostics;
 using System.Drawing;
-using TRDataControl.Remapping;
 using TRImageControl.Packing;
 using TRLevelControl.Model;
