+++ TRDataControl/Transport/TR2/TR2DataImporter.cs
@@ -1,5 +1,4 @@
-﻿using TRDataControl.Remapping;
-using TRImageControl;
+﻿using TRImageControl;
 using TRImageControl.Packing;
 using TRLevelControl.Model;
