+++ TRDataControl/Transport/TR3/TR3TextureRemapper.cs
@@ -1,9 +1,9 @@
 ﻿using TRImageControl.Packing;
 using TRLevelControl.Model;
 
-namespace TRDataControl.Remapping;
+namespace TRDataControl;
 
-public class TR5TextureRemapper : TRTextureRemapper<TR5Level>
+public class TR3TextureRemapper : TRTextureRemapper<TR3Level>
 {
     public override List<TRAnimatedTexture> AnimatedTextures
         => _level.AnimatedTextures;
@@ -17,5 +17,5 @@ public class TR5TextureRemapper : TRTextureRemapper<TR5Level>
         .Concat(_level.DistinctMeshes.SelectMany(m => m.TexturedFaces));
 
     protected override TRTexturePacker CreatePacker()
-        => new TR5TexturePacker(_level, TRGroupPackingMode.Object, 32);
+        => new TR3TexturePacker(_level, 32);
 }
