+++ TRDataControl/Transport/TR4/TR4DataImporter.cs
@@ -1,5 +1,4 @@
-﻿using TRDataControl.Remapping;
-using TRImageControl.Packing;
+﻿using TRImageControl.Packing;
 using TRLevelControl.Model;
 
 namespace TRDataControl;
