+++ TRDataControl/Transport/TRDataImporter.cs
@@ -1,7 +1,6 @@
 ﻿using RectanglePacker.Events;
 using TRImageControl.Packing;
 using TRLevelControl.Model;
-using TRModelTransporter.Transport;
 
 namespace TRDataControl;
