+++ TRDataControl/Transport/TRDataTransport.cs
@@ -1,5 +1,4 @@
-﻿using TRDataControl.Remapping;
-using TRImageControl.Packing;
+﻿using TRImageControl.Packing;
 using TRLevelControl.Model;
 
 namespace TRDataControl;
