+++ TRDataControl/Utilities/Deduplication/TR2LevelTextureDeduplicator.cs
@@ -1,7 +1,7 @@
 ﻿using Newtonsoft.Json;
+using TRDataControl;
 using TRImageControl.Packing;
 using TRLevelControl.Model;
-using TRModelTransporter.Helpers;
 
 namespace TRModelTransporter.Utilities;
