+++ TRRandomizerCore/Randomizers/TR1/TR1OutfitRandomizer.cs
@@ -8,7 +8,6 @@ using TRImageControl.Textures;
 using TRLevelControl;
 using TRLevelControl.Helpers;
 using TRLevelControl.Model;
-using TRModelTransporter.Helpers;
 using TRRandomizerCore.Helpers;
 using TRRandomizerCore.Levels;
 using TRRandomizerCore.Meshes;
