+++ TRRandomizerCore/Textures/Wireframing/AbstractTRWireframer.cs
@@ -1,10 +1,10 @@
 using RectanglePacker.Organisation;
 using System.Drawing;
 using System.Drawing.Drawing2D;
+using TRDataControl;
 using TRImageControl;
 using TRImageControl.Packing;
 using TRLevelControl.Model;
-using TRModelTransporter.Helpers;
 
 namespace TRRandomizerCore.Textures;
