+++ TRRandomizerCore/Textures/Wireframing/TR2Wireframer.cs
@@ -1,9 +1,9 @@
 ﻿using System.Drawing;
+using TRDataControl;
 using TRImageControl;
 using TRImageControl.Packing;
 using TRLevelControl.Helpers;
 using TRLevelControl.Model;
-using TRModelTransporter.Helpers;
 using TRRandomizerCore.Utilities;
 
 namespace TRRandomizerCore.Textures;
