+++ TRRandomizerCore/Textures/Wireframing/TR3Wireframer.cs
@@ -1,10 +1,10 @@
 ﻿using System.Drawing;
 using System.Drawing.Drawing2D;
+using TRDataControl;
 using TRImageControl;
 using TRImageControl.Packing;
 using TRLevelControl.Helpers;
 using TRLevelControl.Model;
-using TRModelTransporter.Helpers;
 using TRRandomizerCore.Utilities;
 
 namespace TRRandomizerCore.Textures;
